Move exceptions into single module
Also removes some unused exceptions

Fixes #150
