--- pypump/exception/ServerError.py
@@ -1,25 +1,0 @@
-##
-#   Copyright (C) 2013 Jessica T. (Tsyesika) <xray7224@googlemail.com>
-#
-#   This program is free software: you can redistribute it and/or modify
-#   it under the terms of the GNU General Public License as published by
-#   the Free Software Foundation, either version 3 of the License, or
-#   (at your option) any later version.
-#
-#   This program is distributed in the hope that it will be useful,
-#   but WITHOUT ANY WARRANTY; without even the implied warranty of
-#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-#   GNU General Public License for more details.
-#
-#   You should have received a copy of the GNU General Public License
-#   along with this program. If not, see <http://www.gnu.org/licenses/>.
-##
-
-import json
-
-
-class ServerError(Exception):
-
-    def __init__(self, data, *args, **kwargs):
-        data = json.loads(data)
-        super(ServerError, self).__init__(data["error"], *args, **kwargs)
