+++ pypump/models/person.py
@@ -18,7 +18,7 @@
 import six
 
 from pypump.models import PumpObject, Addressable
-from pypump.exception import PyPumpException
+from pypump.exceptions import PyPumpException
 from pypump.models.feed import (Followers, Following, Lists,
                                 Favorites, Inbox, Outbox)
